+++ src/tests/loader-cleanup-test.c
@@ -176,7 +176,7 @@ static size_t bad_fwrite (const void *ptr, size_t size, size_t nmemb, FILE *stre
     if (fwrite_count_trigger == fwrite_count || event_count == event_count_trigger) {
 	error_injected++;
 	errno = ENOSPC;
-	r = -1;
+	r = (size_t) -1;
     } else {
 	r = fwrite(ptr, size, nmemb, stream);
 	if (r!=nmemb) {
